Reorganize io modules
